# core/api/openapi_v1.yaml
openapi: 3.0.3
info:
  title: Wachbuch API
  description: |
    Versionierte JSON-API für selbst gehostetes Wachbuch.
    Widerrufbare App-Tokens (`wb_…`) und stationsbezogene Rechte.
    Wachalltag umfasst Schichtübergaben, Mängel, Fahrzeug-/Gerätestatus,
    Checklisten, Schlüssel-/Pool-Geräte, Quittierungen und leichte Auswertung.
    Zusätzliche, je Station aktivierbare Module: Pinnwand (`/pinnwand/`) sowie
    der Ende-zu-Ende-verschlüsselte Nachrichtenbereich aus Wachenchat,
    privaten Zweiergesprächen, Gruppenräumen und Secure Mail
    (`/chat/*`, `/post/*`). Der Server speichert und liefert dabei nur
    Ciphertext und Schlüsselumschläge aus; er kann Nachrichteninhalte nicht lesen.
    Patienten-, Einsatz-, Alarmierungs- und Vorgangsdaten sind ausdrücklich
    nicht Teil dieser API.
  version: "1.4.0"
servers:
  - url: /api/v1
    description: Relative to the Wachbuch origin
security:
  - TokenAuth: []
paths:
...
  /handovers/{id}/:
    get:
      summary: Handover detail
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      responses:
        "200": { description: Handover detail }
        "404": { $ref: "#/components/responses/Error" }
...
  /handovers/{id}/ack/:
    post:
      summary: Acknowledge the explicitly read revision; idempotent per user and revision
      parameters: [{$ref: "#/components/parameters/IntegerId"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Existing acknowledgement returned unchanged
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      ok: { type: boolean }
                  - { $ref: "#/components/schemas/HandoverAck" }
        "201":
          description: Acknowledgement created
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      ok: { type: boolean }
                  - { $ref: "#/components/schemas/HandoverAck" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: Token <secret> or Bearer <secret>"
  parameters:
    IntegerId:
      in: path
      name: id
      required: true
      schema: { type: integer }
  responses:
    Error:
      description: Canonical API error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code:
              type: string
              enum: [validation_error, auth_required, forbidden, mfa_required, mfa_setup_required, not_found, rate_limit, conflict, server_error]
            message: { type: string }
            correlation_id: { type: string }
...
    HandoverAck:
      type: object
      required: [handover_id, by, at]
      properties:
        handover_id: { type: integer }
        by: { type: string }
        at: { type: string, format: date-time }
        version: { type: integer, nullable: true }
...